Refactor: remove column definition
tinyint가 적용되지 않아 삭제함
