+++ src/main/java/beforespring/yourfood/app/member/domain/MemberRepository.java
@@ -1,4 +1,4 @@
-package beforespring.yourfood.member.domain;
+package beforespring.yourfood.app.member.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
