+++ src/main/java/beforespring/yourfood/app/restaurant/domain/CuisineType.java
@@ -1,4 +1,4 @@
-package beforespring.yourfood.restaurant.domain;
+package beforespring.yourfood.app.restaurant.domain;
 
 public enum CuisineType {
     KOREAN,
