+++ src/main/java/beforespring/yourfood/app/restaurant/domain/RestaurantRepository.java
@@ -1,7 +1,6 @@
-package beforespring.yourfood.restaurant.domain;
+package beforespring.yourfood.app.restaurant.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 public interface RestaurantRepository extends JpaRepository<Restaurant, Long> {
 }
