+++ src/main/java/beforespring/yourfood/app/review/domain/Review.java
@@ -1,11 +1,14 @@
-package beforespring.yourfood.review;
+package beforespring.yourfood.app.review.domain;
 
-import beforespring.yourfood.member.domain.Member;
-import beforespring.yourfood.restaurant.domain.Restaurant;
+import beforespring.yourfood.app.member.domain.Member;
+import beforespring.yourfood.app.restaurant.domain.Restaurant;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
@@ -13,6 +16,7 @@ import java.time.LocalDateTime;
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
+@EntityListeners(AuditingEntityListener.class)
 public class Review {
     @Id
     @Column(name = "review_id")
@@ -21,21 +25,22 @@ public class Review {
 
     @JoinColumn(name = "member_id", nullable = false)
     @ManyToOne(fetch = FetchType.LAZY)
-    private Member member;
+    private Member member; //추후 memberId만 참조하는 것으로 변경
 
     @JoinColumn(name = "restaurant_id", nullable = false)
     @ManyToOne(fetch = FetchType.LAZY)
     private Restaurant restaurant;
 
-    @Column(nullable = false)
     private String content;
 
     @Column(columnDefinition = "TINYINT", nullable = false)
-    private Byte rating;
+    private Integer rating;
 
+    @CreatedDate
     @Column(name = "created_at", nullable = false)
     private LocalDateTime createdAt;
 
+    @LastModifiedDate
     @Column(name = "updated_at")
     private LocalDateTime updatedAt;
 
@@ -45,28 +50,21 @@ public class Review {
         Member member,
         Restaurant restaurant,
         String content,
-        Byte rating,
-        LocalDateTime createdAt,
-        LocalDateTime updatedAt) {
+        Integer rating){
         this.member = member;
         this.restaurant = restaurant;
         this.content = content;
         this.rating = rating;
-        this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
     }
 
     /**
      * 리뷰 업데이트
      *
-     * @param content   수정할 내용
-     * @param rating    수정할 평점
-     * @param updatedAt 수정 시간
+     * @param content 수정할 내용
+     * @param rating  수정할 평점
      */
-    public void updateReview(String content, Byte rating, LocalDateTime updatedAt) {
+    public void updateReview(String content, Integer rating) {
         this.content = content;
         this.rating = rating;
-        this.updatedAt = updatedAt;
     }
-
 }
