+++ src/main/java/beforespring/yourfood/app/review/domain/ReviewRepository.java
@@ -1,4 +1,4 @@
-package beforespring.yourfood.review;
+package beforespring.yourfood.app.review.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
