+++ src/main/java/beforespring/yourfood/auth/authmember/domain/AuthMemberRepository.java
@@ -4,7 +4,6 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-@Repository
 public interface AuthMemberRepository extends JpaRepository<AuthMember, Long> {
     Optional<AuthMember> findByUsername(String username);
 }
