+++ src/main/java/beforespring/yourfood/auth/authmember/domain/ConfirmRepository.java
@@ -4,7 +4,6 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-@Repository
 public interface ConfirmRepository extends JpaRepository<Confirm, Long> {
-    Optional<Confirm> findByMember(AuthMember authMember);
+    Optional<Confirm> findByAuthMember(AuthMember authMember);
 }
